refactor: CRUD for task status entities

+++ src/main/java/hexlet/code/controller/api/TaskStatusesController.java
@@ -1,10 +1,12 @@
 package hexlet.code.controller.api;
 
-import hexlet.code.dto.task.status.TaskStatusCreateDTO;
-import hexlet.code.dto.task.status.TaskStatusDTO;
-import hexlet.code.dto.task.status.TaskStatusUpdateDTO;
+import hexlet.code.dto.taskstatus.TaskStatusCreateDTO;
+import hexlet.code.dto.taskstatus.TaskStatusDTO;
+import hexlet.code.dto.taskstatus.TaskStatusUpdateDTO;
+import hexlet.code.exception.ResourceDestroyNotAllowedException;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.TaskStatusMapper;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.TaskStatusRepository;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,12 +34,17 @@ public class TaskStatusesController {
     @Autowired
     private TaskStatusMapper taskStatusMapper;
 
+    @Autowired
+    private TaskRepository taskRepository;
+
     @GetMapping("")
     public ResponseEntity<List<TaskStatusDTO>> index() {
         var taskStatuses = taskStatusRepository.findAll();
+
         var result = taskStatuses.stream()
                 .map(taskStatusMapper::map)
                 .toList();
+
         return ResponseEntity.ok()
                 .header("X-Total-Count", String.valueOf(result.size()))
                 .body(result);
@@ -48,6 +55,7 @@ public class TaskStatusesController {
     public TaskStatusDTO show(@PathVariable Long id) {
         var taskStatus = taskStatusRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Task status with id " + id + " not found"));
+
         return taskStatusMapper.map(taskStatus);
     }
 
@@ -55,7 +63,9 @@ public class TaskStatusesController {
     @ResponseStatus(HttpStatus.CREATED)
     public TaskStatusDTO create(@Valid @RequestBody TaskStatusCreateDTO taskStatusData) {
         var taskStatus = taskStatusMapper.map(taskStatusData);
+
         taskStatusRepository.save(taskStatus);
+
         return taskStatusMapper.map(taskStatus);
     }
 
@@ -64,14 +74,23 @@ public class TaskStatusesController {
     public TaskStatusDTO update(@Valid @RequestBody TaskStatusUpdateDTO taskStatusData, @PathVariable Long id) {
         var taskStatus = taskStatusRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Task status with id " + id + " not found"));
+
         taskStatusMapper.update(taskStatusData, taskStatus);
         taskStatusRepository.save(taskStatus);
+
         return taskStatusMapper.map(taskStatus);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void destroy(@PathVariable Long id) {
-        taskStatusRepository.deleteById(id);
+        if (!taskStatusRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Task status with id " + id + " not found");
+        }
+        if (!taskRepository.existsByTaskStatusId(id)) {
+            taskStatusRepository.deleteById(id);
+        } else {
+            throw new ResourceDestroyNotAllowedException("Cannot delete task status: related task exist");
+        }
     }
 }
